Moved StorageFileWriter into SQLiteStorageFileWriter
